+++ tests/xyz/oihana/schema/constants/traits/thesaurus/DeliveryMethodTermTraitTest.php
@@ -12,6 +12,7 @@ class DeliveryMethodTermTraitTest extends TestCase
     {
         $host = new class { use DeliveryMethodTermTrait; };
 
+        $this->assertSame( 'chargeTiming'          , $host::CHARGE_TIMING );
         $this->assertSame( 'freeShippingThreshold' , $host::FREE_SHIPPING_THRESHOLD );
         $this->assertSame( 'shippingRate'          , $host::SHIPPING_RATE );
         $this->assertSame( 'vat'                   , $host::VAT );